Refactor: Move AppState and AppDatabase to data package
